feat: refactor into teamclient
