+++ src/test/java/com/spotify/github/v3/clients/OrganisationClientTest.java
@@ -69,68 +69,4 @@ public class OrganisationClientTest {
     json = Json.create();
     when(github.json()).thenReturn(json);
   }
-
-  @Test
-  public void getTeam() throws Exception {
-    final CompletableFuture<Team> fixture =
-        completedFuture(json.fromJson(getFixture("team_get.json"), Team.class));
-    when(github.request("/orgs/github/teams/justice-league", Team.class)).thenReturn(fixture);
-    final Team team = organisationClient.getTeam("justice-league").get();
-    assertThat(team.id(), is(1));
-    assertThat(team.name(), is("Justice League"));
-  }
-
-  @Test
-  public void listTeams() throws Exception {
-    final CompletableFuture<List<Team>> fixture =
-        completedFuture(json.fromJson(getFixture("teams_list.json"), LIST_TEAMS));
-    when(github.request("/orgs/github/teams", LIST_TEAMS)).thenReturn(fixture);
-    final List<Team> teams = organisationClient.listTeams().get();
-    assertThat(teams.get(0).slug(), is("justice-league"));
-    assertThat(teams.get(1).slug(), is("x-men"));
-    assertThat(teams.size(), is(2));
-  }
-
-  @Test
-  public void deleteTeam() throws Exception {
-    final CompletableFuture<Response> response = completedFuture(mock(Response.class));
-    final ArgumentCaptor<String> capture = ArgumentCaptor.forClass(String.class);
-    when(github.delete(capture.capture())).thenReturn(response);
-
-    CompletableFuture<Void> deleteResponse = organisationClient.deleteTeam("justice-league");
-    deleteResponse.get();
-    assertThat(capture.getValue(), is("/orgs/github/teams/justice-league"));
-  }
-
-  @Test
-  public void createTeam() throws Exception {
-    final TeamCreate teamCreateRequest =
-        json.fromJson(
-            getFixture("teams_request.json"),
-            TeamCreate.class);
-
-    final CompletableFuture<Team> fixtureResponse = completedFuture(json.fromJson(
-        getFixture("team_get.json"),
-        Team.class));
-    when(github.post(any(), any(), eq(Team.class))).thenReturn(fixtureResponse);
-    final CompletableFuture<Team> actualResponse = organisationClient.createTeam(teamCreateRequest);
-
-    assertThat(actualResponse.get().name(), is("Justice League"));
-  }
-
-  @Test
-  public void updateTeam() throws Exception {
-    final TeamCreate teamCreateRequest =
-        json.fromJson(
-            getFixture("teams_patch.json"),
-            TeamCreate.class);
-
-    final CompletableFuture<Team> fixtureResponse = completedFuture(json.fromJson(
-        getFixture("teams_patch_response.json"),
-        Team.class));
-    when(github.patch(any(), any(), eq(Team.class))).thenReturn(fixtureResponse);
-    final CompletableFuture<Team> actualResponse = organisationClient.updateTeam(teamCreateRequest, "justice-league");
-
-    assertThat(actualResponse.get().name(), is("Justice League2"));
-  }
 }
